Fix all contract tests aligned to open specification
