+++ src/main/kotlin/com/store/exception/ProductDetailsException.kt
@@ -1,10 +1,11 @@
 package com.store.exception
 
+import org.springframework.http.HttpStatus
+import org.springframework.web.bind.annotation.ResponseStatus
 import java.time.LocalDateTime
 
+@ResponseStatus(HttpStatus.BAD_REQUEST)
 data class ProductDetailsException(
         val timeStamp: LocalDateTime = LocalDateTime.now(),
-        val status: Int,
-        val error: String,
-        val path: String
-)
+        override val message: String
+) : Exception( message)
